Added calcTime on rowName input
seems dumb, but without it someone could add a row name after calcTime has run on the time inputs and then ask for the CSV and the rowName wouldn't be in it.  Also note that form validation is being weird - can't type non-numbers at all so no errors, just nothing happens.